feat: mouv auth in module and fusion user crud and user auth
