+++ src/app.module.ts
@@ -1,9 +1,8 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 import { PrismaModule } from './lib/prisma/prisma.module';
-import { UserModule } from './user/user.module';
-import { AuthModule } from './auth/auth.module';
-import { UserModule as CrudUserModule } from './modules/users/user.module';
+import { UserModule } from './modules/users/user.module';
+import { AuthModule } from './modules/auth/auth.module';
 
 @Module({
   imports: [
@@ -11,7 +10,6 @@ import { UserModule as CrudUserModule } from './modules/users/user.module';
     PrismaModule,
     UserModule,
     AuthModule,
-    CrudUserModule,
   ],
   controllers: [],
   providers: [],
