--- src/user/user.module.ts
@@ -1,12 +1,0 @@
-import { Module } from '@nestjs/common';
-import { UserController } from './user.controller';
-import { UserService } from './user.service';
-import { PrismaModule } from 'src/lib/prisma/prisma.module';
-
-@Module({
-  imports: [PrismaModule],
-  controllers: [UserController],
-  providers: [UserService],
-  exports: [UserService],
-})
-export class UserModule {}
